ci: split the benchmark into a ci/ folder
Move the constants, helpers, corpus, and benchmark functions into small
files under bench/src/ci/, leaving ci.ts as a thin entry that parses args
and writes the result.
