+++ bench/src/ci/constants.ts
@@ -1,0 +1,30 @@
+export const importMap = {
+  css: ['@panda/css'],
+  recipe: ['@panda/recipes'],
+  pattern: ['@panda/patterns'],
+  jsx: ['@panda/jsx'],
+  tokens: ['@panda/tokens'],
+}
+
+export const LAYERS = ['reset', 'base', 'tokens', 'recipes', 'utilities'] as const
+
+export const SPACING_VALUES = 133
+
+export const SPACING_PROPERTIES = [
+  'padding',
+  'paddingTop',
+  'paddingBottom',
+  'paddingLeft',
+  'paddingRight',
+  'paddingX',
+  'paddingY',
+  'margin',
+  'marginTop',
+  'marginBottom',
+  'marginLeft',
+  'marginRight',
+  'marginX',
+  'marginY',
+]
+
+export const CONTAINERS = 14
